fix(ci): 修复 Linux GCC -Werror 编译失败（Performance Tracking/Nightly 流水线）
- MemoryInspectionAPI: inspectHeap switch 补全 7 个新增 ValueType（-Werror=switch），
  valueTypeName/heapStructNameForType 同步补齐映射
- Interpreter: numericBinaryOp 移除参数 return 的冗余 std::move（-Werror=redundant-move）
- InterpreterCalls: constructClassInstance 参数数量比较消除符号不匹配（-Werror=sign-compare）
- TypeChecker: TypeInfo 添加显式构造函数，消除部分聚合初始化警告（-Werror=missing-field-initializers）
- CrashHandler: POSIX 路径拼接缓冲扩容为 dir+fname 最大长度和（-Werror=format-truncation）

本地验证: MSVC Debug 构建零错误, ctest 4076/4076 全绿

+++ interpreter/InterpreterCalls.cpp
@@ -646,8 +646,8 @@ Value Interpreter::constructClassInstance(FunCall& node) {
     FunDecl* initMethod = findMethod(*cls, "init");
 
     // P0-2 fix: 支持默认参数，参数数量可在 [requiredParamCount, params.size()] 范围内
-    if (initMethod &&
-        (node.arguments.size() < initMethod->requiredParamCount || node.arguments.size() > initMethod->params.size())) {
+    if (initMethod && (node.arguments.size() < static_cast<size_t>(initMethod->requiredParamCount) ||
+                       node.arguments.size() > initMethod->params.size())) {
         runtimeError(ErrorFormat::formatStd("构造函数 init 期望 {}-{} 个参数，但传入了 {} 个",
 
                                             initMethod->requiredParamCount, initMethod->params.size(),
